--- models/log.js
@@ -1,20 +1,0 @@
-module.exports = (sequelize, DataTypes) => {
-    const Log = sequelize.define('log', {
-        description: {
-            type: DataTypes.STRING,
-            allowNull: false,
-        },
-        definition: {
-            type: DataTypes.STRING,
-            allowNull: false
-        },
-        result: {
-            type: DataTypes.STRING,
-            allowNull: false
-        },
-        owner: {
-            type: DataTypes.INTEGER
-        }
-    })
-    return Log;
-};
